refactor(cli/kmkc): move purchase common to `common.rs`
