+++ tosho/src/impl/kmkc/purchases.rs
@@ -2,194 +2,12 @@ use color_print::cformat;
 use num_format::{Locale, ToFormattedString};
 use tosho_kmkc::{
     constants::BASE_HOST,
-    models::{EpisodeNode, TicketInfoType, TitleTicketListNode, UserPointResponse},
-    KMClient,
+    models::{EpisodeNode, TicketInfoType},
 };
 
-use crate::{cli::ExitCode, linkify, term::ConsoleChoice};
+use crate::{cli::ExitCode, linkify};
 
-use super::{common::select_single_account, config::Config};
-
-#[derive(Clone)]
-struct PurchasePoint {
-    point: UserPointResponse,
-    ticket: TitleTicketListNode,
-}
-
-async fn common_purchase_select(
-    title_id: i32,
-    account: &Config,
-    console: &crate::term::Terminal,
-) -> (
-    anyhow::Result<Vec<tosho_kmkc::models::EpisodeNode>>,
-    KMClient,
-    Option<PurchasePoint>,
-) {
-    let client = super::common::make_client(&account.into());
-    console.info(&cformat!(
-        "Getting user point for <m,s>{}</>...",
-        account.get_username()
-    ));
-    let user_point = client.get_user_point().await;
-    if let Err(error) = user_point {
-        console.error(&format!("Unable to get user point: {}", error));
-        return (Err(error), client, None);
-    }
-    let user_point = user_point.unwrap();
-
-    console.info(&cformat!(
-        "Getting title information for ID <m,s>{}</>...",
-        title_id
-    ));
-    let results = client.get_titles(vec![title_id]).await;
-    if let Err(error) = results {
-        console.error(&format!("Failed to get title information: {}", error));
-        return (Err(error), client, None);
-    }
-
-    let results = results.unwrap();
-    if results.is_empty() {
-        console.error("Unable to find title information");
-        return (
-            Err(anyhow::anyhow!("Unable to find title information")),
-            client,
-            None,
-        );
-    }
-
-    let result = results.first().unwrap();
-
-    console.info(&cformat!(
-        "Fetching <m,s>{}</> title ticket...",
-        result.title
-    ));
-    let ticket_entry = client.get_title_ticket(result.id).await;
-    if let Err(error) = ticket_entry {
-        console.error(&format!("Failed to get title ticket: {}", error));
-        return (Err(error), client, None);
-    }
-
-    let ticket_entry = ticket_entry.unwrap();
-
-    let mut chapters_entry = vec![];
-    console.info(&cformat!("Fetching <m,s>{}</> chapters...", result.title));
-    for episodes in result.episode_ids.clone().chunks(50) {
-        let chapters = client.get_episodes(episodes.to_vec()).await;
-
-        if let Err(error) = chapters {
-            console.error(&format!("Failed to get chapters: {}", error));
-            return (
-                Err(error),
-                client,
-                Some(PurchasePoint {
-                    point: user_point,
-                    ticket: ticket_entry,
-                }),
-            );
-        }
-
-        chapters_entry.extend(chapters.unwrap());
-    }
-
-    console.info("Your current point balance:");
-    let total_bal = user_point
-        .point
-        .total_point()
-        .to_formatted_string(&Locale::en);
-    let paid_point = user_point.point.paid_point.to_formatted_string(&Locale::en);
-    let free_point = user_point.point.free_point.to_formatted_string(&Locale::en);
-    let premium_ticket = user_point.ticket.total_num.to_formatted_string(&Locale::en);
-    console.info(&cformat!(
-        "  - <bold>Total:</> <cyan!,bold><reverse>{}</>c</cyan!,bold>",
-        total_bal
-    ));
-    console.info(&cformat!(
-        "  - <bold>Paid point:</> <g,bold><reverse>{}</>c</g,bold>",
-        paid_point
-    ));
-    console.info(&cformat!(
-        "  - <bold>Free point:</> <cyan,bold><reverse>{}</>c</cyan,bold>",
-        free_point
-    ));
-    console.info(&cformat!(
-        "  - <bold>Premium ticket:</> <yellow,bold><reverse>{}</> ticket</yellow,bold>",
-        premium_ticket
-    ));
-    console.info(&cformat!(
-        "  - <bold>Title ticket?</bold>: {}",
-        ticket_entry.is_title_available()
-    ));
-
-    console.info("Title information:");
-    console.info(&cformat!("  - <bold>ID:</> {}", result.id));
-    console.info(&cformat!("  - <bold>Title:</> {}", result.title));
-    console.info(&cformat!(
-        "  - <bold>Chapters:</> {} chapters",
-        chapters_entry.len()
-    ));
-
-    // Only show unpaid chapters
-    let unpaid_chapters: Vec<&tosho_kmkc::models::EpisodeNode> = chapters_entry
-        .iter()
-        .filter(|ch| !ch.is_available())
-        .collect();
-
-    let select_choices: Vec<ConsoleChoice> = unpaid_chapters
-        .iter()
-        .map(|ch| {
-            if ch.is_ticketable() {
-                ConsoleChoice {
-                    name: ch.id.to_string(),
-                    value: format!("{} ({}P/Ticket)", ch.title, ch.point),
-                }
-            } else {
-                ConsoleChoice {
-                    name: ch.id.to_string(),
-                    value: format!("{} ({}P)", ch.title, ch.point),
-                }
-            }
-        })
-        .collect();
-
-    let selected_chapters = console.select("Select chapter to purchase", select_choices);
-    match selected_chapters {
-        Some(selected) => {
-            let mapped_chapters: Vec<tosho_kmkc::models::EpisodeNode> = selected
-                .iter()
-                .map(|ch| {
-                    let ch_id = ch.name.parse::<i32>().unwrap();
-                    let ch = chapters_entry
-                        .iter()
-                        .find(|ch| ch.id == ch_id)
-                        .unwrap()
-                        .clone();
-
-                    ch
-                })
-                .collect();
-
-            (
-                Ok(mapped_chapters),
-                client,
-                Some(PurchasePoint {
-                    point: user_point,
-                    ticket: ticket_entry,
-                }),
-            )
-        }
-        None => {
-            console.warn("Aborted!");
-            (
-                Ok(vec![]),
-                client,
-                Some(PurchasePoint {
-                    point: user_point,
-                    ticket: ticket_entry,
-                }),
-            )
-        }
-    }
-}
+use super::common::{common_purchase_select, select_single_account};
 
 pub(crate) async fn kmkc_purchase(
     title_id: i32,
@@ -204,7 +22,8 @@ pub(crate) async fn kmkc_purchase(
     }
 
     let account = account.unwrap();
-    let (results, client, user_point) = common_purchase_select(title_id, &account, console).await;
+    let (results, _, client, user_point) =
+        common_purchase_select(title_id, &account, false, false, console).await;
 
     match (results, user_point) {
         (Ok(results), Some(user_point)) => {
@@ -389,7 +208,8 @@ pub(crate) async fn kmkc_purchase_precalculate(
     }
 
     let account = account.unwrap();
-    let (results, _, user_point) = common_purchase_select(title_id, &account, console).await;
+    let (results, _, _, user_point) =
+        common_purchase_select(title_id, &account, false, false, console).await;
 
     match (results, user_point) {
         (Ok(results), Some(user_point)) => {
